Use sc1.ini/dmd1.conf config file if found
Using D2 and D1 (specially with Tango) in the same computer can be
challenging, because both share the same config file name. To ease the
transition, D1 now looks for a file called sc1.ini (Windows) or
dmd1.conf (Posix) first (in the same locations as usual) and only if it
doesn't find it, it looks for the old standard name without the version
number.

This change is somehow backwards compatible (it will only break if for
some reason you already had a file with that name) and should be
reasonably harmless.