gc: Rename gcalloc.d to os.d
In the future this file will hold more OS-specific functions needed by
the GC, and this file looks like a good place to put them, so all
OS-specific functions are grouped in a single module.

Also, having the prefix gc on every file, when all this is inside a gc
package, doesn't seem to make a lot of sense, thus it will be removed
progressively.