fix Issue 10684 - Refused array op with array literal
Accept array literals as the operand of array-ops.

---

Since, array-ops had required explicit slice to the array literal operands,
eg. `a[] = [1,2][] + c[];`

By the commit:
https://github.com/9rnsr/dmd/commit/1da1f38f52e90547afb93a0bb16c01ca7b66a93f
now the redundant slice operators at the right side of an array literal is
entirely removed - `[1,2,3][][]` is rewritten to `[1,2,3]`.

Although it has broke existing code using array-ops, I still cannot think
the optimization was incorrect.

On the other hand, I could not find the description about the necessity of
explicit slice on array literal from the language spec.
http://dlang.org/arrays.html#array-operations
It looked to me that is just an implementation-specific restriction.

So, I implemented this as the easiest way to fix the issue.