Split mangle test into two files
We currently have two VarDeclarations in mangle.d which reference
the same symbol. This causes problems in the GCC backend as it does
not realize it actually deals with the same symbol and therefore
reorders reads and writes. It's not clear if this will be fixed in the
GCC backend at all as the fix would break other stuff. It's also
kind of weird to reference a symbol in this way. The correct way
is to use aliases.
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=60214

Anyway, this test was likely not supposed to test if the compiler
recognizes that these are the same variables, it was probably just to
test that we can read/write to a symbol with a specific mangle.