Use single test runner executable for unit tests.
This is not only much cleaner than the old hack that depended
on the tested modules not being linked in a second time from
the druntime/Phobos static libraries, but the hack obviously
also doesn't work for shared libraries.

Also unify the naming scheme and adapt the Travis config to
respect it.