driver: Emit __bss_start/_end as weak symbols to work around linker bug.
A similar problem as in upstream issue 13025 has been
observerd for LDC on Arch Linux (GNU ld 2.24) as well. This
is essentially the same workaround as Martin came up with in
D-Programming-Language/druntime#889. Because the linker
defines the symbols anyway as part of the default linker
script, the weak linkage actually makes no difference.