Fix bad assumption about clock speed in MonoTime unit tests.
This code worked just fine so long as the two subsequent calls to
MonoTime.currTime returned different results, but apparently, some
machines, CPUs are fast enough in comparison to their system clock that
that's not guaranteed, so the diff was 0 instead of at least 1, and the
test failed. And looking at it, it probably should have been >= rather
than > anyway (at least based on the variable name - min).

This should make it so that the tests stop failing on Daniel Murphy's
machine.