CMake: Remove unsupported, untested USE_BOHEM_GC option.
I don't think anyone ever tried to use it recently. Among other
problems, it expected "libgc.a" to be present in the source
direcotry.

In the unlikely case that somebody wants to experiment with the
Boehm GC again, the required flags are trivial to add via
CMAKE_CXX_FLAGS, etc.