[cleanup] Remove LDC modifications from mostly unused front-end inliner.
The inliner is also used for default argument handling, so we
need to keep it around, but we no longer need any of the cost
tweaks. In fact, this could even fix some potential bugs with
erroneously accepted default arguments.

We will likely want to bring something like InlineCostVisitor
back for cross-module inlining, but with visitors we can do
this without editing the source now.