Fix for issue #624.
In case of a ctor 'this' is passed to the function and is also the
return value. This is the perfect case for the 'Returned' attribute.
But the x86_64 ABI missed the fact that the return value was a pointer
to a struct and did a struct rewrite, resulting in an bitcast
incompatible type.