Add support for compiling Objective-C code as extra sources.
Add the `EXTRA_OBJC_SOURCES` variable and a list of Objective-C files
to compile in a comment in the test file.

Test files with the `EXTRA_OBJC_SOURCES` variable will be ignored
unless the `D_OBJC` environment variable is set to `1`. This gives a way
to only compile D/Objective-C code on platforms where it's supported.