Make WANTvalue default
The difference 0 and WANTvalue was used in AndAndExp/OrOrExp to make error for the void expressions if value is required.

But today, the appearance of invalid void sub-expression is structly checked by Expresssion::rvalue() etc in each semantic() calls, so the check in optimize() is unnecessary.

Therefore, change the value of WANTvalue to 0 and make that the default behavior of optimize().