Fix for issue #874.
After running `git gc` cmake may fail to configure the build.
`git gc` deleted the file `.git/refs/head/merge-2.067` (my current branch) and
moved the reference into file `.git/packed-refs`. This file is not parsed for
the reference.

Solution is (a) parse `.git/packed-refs` if ref is not found or (b) use

    git rev-parse --symbolic-full-name HEAD

to get the reference and

    git show-ref --hash ${HEAD_REF}

to get the hash.

This commits uses solution (a).