Fix for issue #855.
If several modules are linked together, an empty module is used as
target. Starting with LLVM 3.6, there seems to be a slight difference
between an empty module and a module with IR.
The PR uses always the first module as the target. This has the nice
side effect that we are no longer affected by LLVM bug 11479.