Implement cent/ucent data types.
This is a first try at implementing the data types cent and ucent.
All changes are wrapped in #if WANT_CENT .. #endif. It takes
advantage of the GCC builtin type __int128 and is currently only
enabled if compiled with gcc.

As soon as this is stable it should go upstream, too.