Always link libcurl when Phobos is linked
This is the only sensible solution until e.g.
D-Programming-Language/phobos#3009 is merged, which makes
Phobos load libcurl dynamically. The issue with the current
situation is is that the user cannot even manually link
libcurl if linker errors occur, because it will appear
before libphobos2-ldc in the linker command line.

GitHub: Fixes #906.