Fix test code bug for the FinalizeError.
The old code had accidentally worked. There was two `delete m`s in the try
block - explicit one and implcit for the scope destruction. However, the
second `delete` was never invoked until now, because:

a. `DeleteExp::canThrow()` was not overloaded so it had returned false always.
b. `TryFinallyStatement::semantic()` has an optimization if no exception
   will be thrown from the try block, it's translated to CompoundStatements.

By that, the scope destruction had not been placed in finalize block,
then it had just been skipped when the first `delete` throws `FinalizeError`.