LLVM 3.7+: fix toConstElem(SymOffExp*).
The new type parameter for `llvm::ConstantExpr::getGetElementPtr()` is
checked for being the pointee of the provided base (or more precisely,
the pointee of the base's scalar type).
So for the 'ugly' GEP, i8* as type for an i8* base was clearly wrong,
and the 'nice' GEP with the fixed i8 type would only work for i8* bases.

This fixes dmd-testsuite issues on Win64, e.g., runnable/test11.d.