Elide codegen if TypeInfo is speculative
Even if a struct is defined as non-root symbol, some built-in operations
request its TypeInfo. For those, today TypeInfo_Struct is generated in COMDAT.

On the other hand, TypeInfoDeclaration is always generated when the TypeInfo is requested,
even if the code exists in speculative scope. By that, some unneeded
TypeInfo objects had excessively generated.
If a TypeInfo is not actually used, we should stop its generation.

To satisfy two requirements, I added "speculative TypeInfo" concept
associated with template instances.

By this change, when a TypeInfo is *requested*, the codegen pass still
visit it always. But if the corresponding type is instantiated struct or
its derived, and is definitely speculative, the TypeInfo code also won't
be emit into object file.