Appveyor: use a pre-built LLVM with proper debug infos
The archive is 170 MB now (uncompressed: 2.3 GB), so it cannot
be cached by Appveyor anymore. Anyway, the download is quick, and
Dropbox should be able to handle the traffic. ;)

This version has been compiled with explicit C and C++ flags for
RelWithDebInfo, by replacing /Zi with /Z7, to embed the debug infos
in the .obj files directly. Otherwise, the separate .pdb files
aren't installed to the LLVM installation folder and were thus
missing in the previous version.