Use a more robust way of dumping a LL value to memory.
This fixes runnable\sdtor.d for Win64 (previous ICE in test11763()).
The temporary is rewritten to `(auto x = HasDtortest11763(), x)`,
and `x` is somehow already an alloca (most likely due it being an
instance of an empty struct with a dummy size of 1 byte).

Not sure if this is enough or whether we should actually make sure
these empty structs get represented by a dummy i8 value. Ideally,
empty structs wouldn't require any memory at all.