Remove global.gag check in DsymbolExp.resolve()
When an instance that will be converted to a constant exists,
the instance representation "foo!tiargs" is treated like a
variable name, and its recursive appearance check (note that
it's equivalent with a recursive instantiation of foo) is done
separately from the circular initialization check for the
eponymous enum variable declaration.

By the approach, we can remove problematic `global.gag` check in
`DsymbolExp::semantic()` without breaking `fail_compilation/diag10141.d`.