Don't try to find a program via environment variable if that is empty (CC)
Whereas most OSs seem to remove the environment variable if it's cleared,
OSX seems to keep it, explicitly cleared so to speak.

I strongly suspect a cleared CC env variable to be the reason for the
triggering assertion when trying to build dmd-testsuite's d_do_test tool
in our OSX CI builds.